% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iLincsCor-pkg.R
\docType{package}
\name{iLincsCor-package}
\alias{iLincsCor}
\alias{iLincsCor-package}
\title{iLincsCor: What the Package Does in One 'Title Case' Line}
\description{
One paragraph description of what the package does as one or more full sentences.
}
\details{
iLincsCor is a package containing iLincsCor class for fast signature 
correlations.
}
