% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iLincsCorExports.R
\name{iLincsCor}
\alias{iLincsCor}
\title{Create a iLincsCor Object from the iLincsCor C++ Class}
\arguments{
\item{prefix}{Path to the library}
}
\value{
A `iLincsCor` object from the _C++_ iLincsCor Class.
}
\description{
Allows for the creation of a iLincsCor Object in _C++_ from _R_
using the _C++_ iLincsCor class.
}
\examples{
##################
## Constructor

# Construct new iLincsCor object called "ben"
lib <- new(iLincsCor, prefix = "/Users/michal/tmp/data/LIB_5/")

##################
## read input vector

v <- lib$read_input("tests/LINCSCP_100.tsv")

## run correlation on 4 CPUs

result <- lib$cor(v,4)

}
